requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai-agents==0.0.15
fastapi-mcp==0.3.4
python-dotenv==1.0.1
